+++ sdk/dotnet/Inputs/NotificationRuleScheduleGetArgs.cs
@@ -15,8 +15,8 @@ namespace Pulumi.Opsgenie.Inputs
         /// <summary>
         /// Name of the notification policy
         /// </summary>
-        [Input("name")]
-        public Input<string>? Name { get; set; }
+        [Input("name", required: true)]
+        public Input<string> Name { get; set; } = null!;
 
         /// <summary>
         /// Kind of matching filter  "match-all", "match-any-condition", "match-all-conditions"