+++ CHANGELOG.md
@@ -2,6 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-* Initial creation of the provider against v0.5.7 of the OpsGenie Terraform Provider
+* Initial creation of the provider against v0.6.2 of the OpsGenie Terraform Provider
 
 ---